==> src/06_scrapePlayerImages.ipynb <==
import pandas as pd
from bs4 import BeautifulSoup
import requests
import time
from selenium import webdriver

team_url = 'http://www.espn.com/nba/players/_/position/pf/league/nba'
request = requests.get(team_url)
soup = BeautifulSoup(request.text, 'html.parser')

[int(x["href"].split("/id/")[1].split("/")[0]) for x in soup.findAll('a') if "player/" in x["href"]]

[x.text.split(", ")[1] + " " + x.text.split(", ")[0] for x in soup.findAll('a') if "player/" in x["href"]]

positions = ["pg", "sg", "sf", "pf", "c"]

playersdf = pd.DataFrame(columns = ["name", "espn_player_id"])

for position in positions:
    team_url = 'http://www.espn.com/nba/players/_/position/' + position + '/league/nba'
    request = requests.get(team_url)
    soup = BeautifulSoup(request.text, 'html.parser') 
    
    ids = [int(x["href"].split("/id/")[1].split("/")[0]) for x in soup.findAll('a') if "player/" in x["href"]]
    names = [x.text.split(", ")[1] + " " + x.text.split(", ")[0] for x in soup.findAll('a') if "player/" in x["href"]]
    
    intermediate = pd.DataFrame({"name" : names,
                                 "espn_player_id" : ids})
    
    playersdf = pd.concat([playersdf, intermediate])

playersdf

playersdf.to_csv("/Users/James/Desktop/Deep_Learning/nba-movement-data/rebounding_data/images/playerspictures.csv",
                 index = False)
